requirements.txt
django-admin-bootstrapped==2.5.0
django-appconf==1.0.1
django-bootstrap-markdown==1.8.1
django-bootstrap-markdown-editor==1.1
django-bootstrap-tags==1.2.0
django-bootstrap3==5.4.0
django-compressor==1.5
django-contrib-comments==1.5.1
django-debug-toolbar==1.3.0
django-imagekit==3.2.6
django-markdown==0.8.4
django-markupfield==1.3.5
django-pagedown==0.1.0
django-password-strength==1.1.1
django-pingback==0.3
django-taggit==0.14.0
django-xmlrpc==0.1.5

colored-traceback==0.2.1
jinja2==2.7.3
nanomsg==1.0
netifaces==0.10.4
ujson==1.33
pymongo==3.0.3
requests==2.7.0

